\name{HPOOBSOLETE}
\alias{HPOOBSOLETE}
\title{Annotation of HPO identifiers by terms defined by Human Phenotype Ontology
    Consortium and their status are obsolete}
\description{
This is an R object mapping HPO identifiers to the specific
terms in defined by Human Phenotype Ontology Consortium and their definition
are obsolete
}
\details{
    All the obsolete HPO terms that are collected in this index will
    no longer exist in other mapping objects.

    Mappings were based on data provided by: Human Phenotype Ontology
    With a date stamp from the source of: 220230405


}
\references{
\url{https://hpo.jax.org/app/data/ontology}
}
\keyword{datasets}

